requirements.txt
flask
Pillow
requests
gunicorn
